Add transaction advisory lock around team deletion

+++ server/commands/teamPermanentDeleter.ts
@@ -19,6 +19,7 @@ import {
   Share,
 } from "@server/models";
 import { sequelize } from "@server/storage/database";
+import { LockHelper } from "@server/storage/LockHelper";
 
 /**
  * Permanently deletes a team and all related data from the database. Note that this does not happen
@@ -109,6 +110,19 @@ async function teamPermanentDeleter(team: Team) {
 
   // Destory team-relation models
   await sequelize.transaction(async (transaction) => {
+    const acquired = await LockHelper.tryAcquire(
+      sequelize,
+      `teamPermanentDeleter:${teamId}`,
+      transaction
+    );
+    if (!acquired) {
+      Logger.info(
+        "commands",
+        `Team ${teamId} is already being destroyed, skipping`
+      );
+      return;
+    }
+
     await AuthenticationProvider.destroy({
       where: {
         teamId,
